# Neptune-respt-api/openapi.yaml
openapi: 3.0.3
info:
  title: Neptune REST API Service
  description: >-
    This is a  API specification for Neptune REST API Service   written in Nodejs
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:8080
...
paths:
...
  /signup:
    post:
      description: ""
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  example: any
                email:
                  example: any
                password:
                  example: any
                pic:
                  example: any
      responses:
        '200':
          description: OK
        '422':
          description: Unprocessable Entity
  /signin:
    post:
      description: ""
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  example: any
                password:
                  example: any
      responses:
        '200':
          description: OK
        '422':
          description: Unprocessable Entity
  /reset-password:
    post:
      description: ""
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  example: any
      responses:
        '200':
          description: OK
        '422':
          description: Unprocessable Entity
  /new-password:
    post:
      description: ""
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  example: any
                token:
                  example: any
      responses:
        '200':
          description: OK
        '422':
          description: Unprocessable Entity
...